# api/swagger.yml
...
consumes:
- application/json
produces:
- application/json
swagger: "2.0"
info:
  description: API documentation
  title: github/chapool/go-wallet
  version: 0.1.0
paths:
...
  /api/v1/wallet/hot-wallet:
    post:
      security:
      - Bearer: []
      description: |-
        Create a hot wallet on the specified chain.
        Only admin users can create hot wallets.
      consumes:
      - application/json
      produces:
      - application/json
      tags:
      - wallet
      summary: Create hot wallet
      operationId: PostCreateHotWalletRoute
      parameters:
      - name: body
        in: body
        required: true
        schema:
          $ref: '#/definitions/postCreateHotWalletPayload'
      responses:
        "200":
          description: Hot wallet created successfully
          schema:
            $ref: '#/definitions/createHotWalletResponse'
        "400":
          description: PublicHTTPValidationError
          schema:
            $ref: '#/definitions/publicHttpValidationError'
        "401":
          description: PublicHTTPError
          schema:
            $ref: '#/definitions/publicHttpError'
        "403":
          description: PublicHTTPError
          schema:
            $ref: '#/definitions/publicHttpError'
        "500":
          description: PublicHTTPError
          schema:
            $ref: '#/definitions/publicHttpError'
...
  /api/v1/wallet/withdraw/{withdrawId}/reject:
    post:
      security:
      - Bearer: []
      description: |-
        Reject a withdraw request.
        Only admin users can reject withdraw requests.
      consumes:
      - application/json
      produces:
      - application/json
      tags:
      - wallet
      summary: Reject withdraw request (Admin only)
      operationId: PostRejectWithdrawRoute
      parameters:
      - type: string
        format: uuid
        description: Withdraw ID to reject
        name: withdrawId
        in: path
        required: true
      - name: body
        in: body
        schema:
          $ref: '#/definitions/postRejectWithdrawPayload'
      responses:
        "200":
          description: Withdraw rejected successfully
          schema:
            $ref: '#/definitions/withdrawResponse'
        "400":
          description: PublicHTTPValidationError
          schema:
            $ref: '#/definitions/publicHttpValidationError'
        "401":
          description: PublicHTTPError
          schema:
            $ref: '#/definitions/publicHttpError'
        "403":
          description: PublicHTTPError
          schema:
            $ref: '#/definitions/publicHttpError'
        "404":
          description: PublicHTTPError
          schema:
            $ref: '#/definitions/publicHttpError'
        "500":
          description: PublicHTTPError
          schema:
            $ref: '#/definitions/publicHttpError'
...
definitions:
...
  createHotWalletResponse:
    type: object
    required:
    - id
    - address
    - chain_type
    - chain_id
    - chain_name
    - derivation_path
    - address_index
    - wallet_type
    - created_at
    properties:
      address:
        type: string
        example: "0x742d35Cc6634C0532925a3b8D4C9db96C4b4d8b6"
      address_index:
        type: integer
        example: 0
      chain_id:
        type: integer
        example: 1
      chain_name:
        type: string
        example: Ethereum Mainnet
      chain_type:
        type: string
        example: evm
      created_at:
        type: string
        format: date-time
        example: "2025-01-01T00:00:00Z"
      derivation_path:
        type: string
        example: m/44'/60'/0'/0/0
      device_name:
        type: string
        example: hot-wallet-1
      id:
        type: string
        format: uuid
        example: 550e8400-e29b-41d4-a716-446655440000
      wallet_type:
        type: string
        example: hot
...
  httpValidationErrorDetail:
    type: object
    required:
    - key
    - in
    - error
    properties:
      error:
        description: Error describing field validation failure
        type: string
      in:
        description: Indicates how the invalid field was provided
        type: string
      key:
        description: Key of field failing validation
        type: string
...
  postCreateHotWalletPayload:
    type: object
    required:
    - chain_id
    - device_name
    properties:
      chain_id:
        description: Chain ID (1 for Ethereum Mainnet, 137 for Polygon, 56 for BSC,
          etc.)
        type: integer
        example: 1
      device_name:
        description: Device name for the hot wallet
        type: string
        example: hot-wallet-1
...
  postRejectWithdrawPayload:
    type: object
    properties:
      reason:
        description: Reason for rejection (optional)
        type: string
        example: Insufficient funds in hot wallet
...
  publicHttpError:
    type: object
    required:
    - status
    - type
    - title
    properties:
      detail:
        description: More detailed, human-readable, optional explanation of the error
        type: string
        example: User is lacking permission to access this resource
      status:
        description: HTTP status code returned for the error
        type: integer
        format: int64
        maximum: 599
        minimum: 100
        x-go-name: Code
        example: 403
      title:
        description: Short, human-readable description of the error
        type: string
        example: Forbidden
      type:
        $ref: '#/definitions/publicHttpErrorType'
  publicHttpErrorType:
    description: Type of error returned, should be used for client-side error handling
    type: string
    enum:
    - generic
    - PUSH_TOKEN_ALREADY_EXISTS
    - OLD_PUSH_TOKEN_NOT_FOUND
    - ZERO_FILE_SIZE
    - USER_DEACTIVATED
    - INVALID_PASSWORD
    - NOT_LOCAL_USER
    - TOKEN_NOT_FOUND
    - TOKEN_EXPIRED
    - USER_ALREADY_EXISTS
    - MALFORMED_TOKEN
    - LAST_AUTHENTICATED_AT_EXCEEDED
    - MISSING_SCOPES
  publicHttpValidationError:
    type: object
    required:
    - validationErrors
    allOf:
    - $ref: '#/definitions/publicHttpError'
    properties:
      validationErrors:
        description: List of errors received while validating payload against schema
        type: array
        items:
          $ref: '#/definitions/httpValidationErrorDetail'
...
  withdrawItem:
    type: object
    required:
    - id
    - user_id
    - to_address
    - token_id
    - amount
    - status
    - created_at
    properties:
      amount:
        type: string
        example: "1.5"
      created_at:
        type: string
        format: date-time
      fee:
        type: string
        example: "0.001"
      id:
        type: string
        format: uuid
      status:
        type: string
        enum:
        - user_withdraw_request
        - signing
        - pending
        - processing
        - confirmed
        - failed
        example: user_withdraw_request
      to_address:
        type: string
        example: "0x742d35Cc6634C0532925a3b8D4C9db96C4b4d8b6"
      token_id:
        type: integer
        example: 1
      tx_hash:
        type: string
        example: 0x...
      user_id:
        type: string
        format: uuid
  withdrawResponse:
    type: object
    required:
    - withdraw
    properties:
      withdraw:
        $ref: '#/definitions/withdrawItem'
...
securityDefinitions:
  Bearer:
    description: |-
      Access token for application access, **must** include "Bearer " prefix.
      Example: `Bearer b4a94a42-3ea2-4af3-9699-8bcbfee6e6d2`
    type: apiKey
    name: Authorization
    in: header
    x-keyPrefix: 'Bearer '
  Management:
    description: Management secret, used for monitoring and infrastructure related
      calls
    type: apiKey
    name: mgmt-secret
    in: query