# internal/types/create_hot_wallet_response.go
// Code generated by go-swagger; DO NOT EDIT.

package types

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// CreateHotWalletResponse create hot wallet response
//
// swagger:model createHotWalletResponse
type CreateHotWalletResponse struct {

	// address
	// Example: 0x742d35Cc6634C0532925a3b8D4C9db96C4b4d8b6
	// Required: true
	Address *string `json:"address"`

	// address index
	// Example: 0
	// Required: true
	AddressIndex *int64 `json:"address_index"`

	// chain id
	// Example: 1
	// Required: true
	ChainID *int64 `json:"chain_id"`

	// chain name
	// Example: Ethereum Mainnet
	// Required: true
	ChainName *string `json:"chain_name"`

	// chain type
	// Example: evm
	// Required: true
	ChainType *string `json:"chain_type"`

	// created at
	// Example: 2025-01-01T00:00:00Z
	// Required: true
	// Format: date-time
	CreatedAt *strfmt.DateTime `json:"created_at"`

	// derivation path
	// Example: m/44'/60'/0'/0/0
	// Required: true
	DerivationPath *string `json:"derivation_path"`

	// device name
	// Example: hot-wallet-1
	DeviceName string `json:"device_name,omitempty"`

	// id
	// Example: 550e8400-e29b-41d4-a716-446655440000
	// Required: true
	// Format: uuid
	ID *strfmt.UUID `json:"id"`

	// wallet type
	// Example: hot
	// Required: true
	WalletType *string `json:"wallet_type"`
}

// Validate validates this create hot wallet response
func (m *CreateHotWalletResponse) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateAddress(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateAddressIndex(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateChainID(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateChainName(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateChainType(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateCreatedAt(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateDerivationPath(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateID(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateWalletType(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *CreateHotWalletResponse) validateAddress(formats strfmt.Registry) error {

	if err := validate.Required("address", "body", m.Address); err != nil {
		return err
	}

	return nil
}

func (m *CreateHotWalletResponse) validateAddressIndex(formats strfmt.Registry) error {

	if err := validate.Required("address_index", "body", m.AddressIndex); err != nil {
		return err
	}

	return nil
}

func (m *CreateHotWalletResponse) validateChainID(formats strfmt.Registry) error {

	if err := validate.Required("chain_id", "body", m.ChainID); err != nil {
		return err
	}

	return nil
}

func (m *CreateHotWalletResponse) validateChainName(formats strfmt.Registry) error {

	if err := validate.Required("chain_name", "body", m.ChainName); err != nil {
		return err
	}

	return nil
}

func (m *CreateHotWalletResponse) validateChainType(formats strfmt.Registry) error {

	if err := validate.Required("chain_type", "body", m.ChainType); err != nil {
		return err
	}

	return nil
}

func (m *CreateHotWalletResponse) validateCreatedAt(formats strfmt.Registry) error {

	if err := validate.Required("created_at", "body", m.CreatedAt); err != nil {
		return err
	}

	if err := validate.FormatOf("created_at", "body", "date-time", m.CreatedAt.String(), formats); err != nil {
		return err
	}

	return nil
}

func (m *CreateHotWalletResponse) validateDerivationPath(formats strfmt.Registry) error {

	if err := validate.Required("derivation_path", "body", m.DerivationPath); err != nil {
		return err
	}

	return nil
}

func (m *CreateHotWalletResponse) validateID(formats strfmt.Registry) error {

	if err := validate.Required("id", "body", m.ID); err != nil {
		return err
	}

	if err := validate.FormatOf("id", "body", "uuid", m.ID.String(), formats); err != nil {
		return err
	}

	return nil
}

func (m *CreateHotWalletResponse) validateWalletType(formats strfmt.Registry) error {

	if err := validate.Required("wallet_type", "body", m.WalletType); err != nil {
		return err
	}

	return nil
}

// ContextValidate validates this create hot wallet response based on context it is used
func (m *CreateHotWalletResponse) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *CreateHotWalletResponse) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *CreateHotWalletResponse) UnmarshalBinary(b []byte) error {
	var res CreateHotWalletResponse
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}